Move subgraph hooks to a subfolder Create a new hook to call get the Sugraph by name
