+++ src/pagePartials/badge/curate/CurateModal.tsx
@@ -24,7 +24,7 @@ import { gradients } from 'thebadge-ui-library'
 import DisplayEvidenceField from '@/src/components/displayEvidence/DisplayEvidenceField'
 import { Address } from '@/src/components/helpers/Address'
 import SafeSuspense from '@/src/components/helpers/SafeSuspense'
-import useBadgeById from '@/src/hooks/useBadgeById'
+import useBadgeById from '@/src/hooks/subgraph/useBadgeById'
 import CurationCriteriaLink from '@/src/pagePartials/badge/curate/CurationCriteriaLink'
 import { RequiredConnection } from '@/src/pagePartials/errors/requiredConnection'
 import { useCurateProvider } from '@/src/providers/curateProvider'
