+++ src/pagePartials/errors/requiredNotHaveBadge.tsx
@@ -4,7 +4,7 @@ import React, { ReactElement } from 'react'
 import { styled } from '@mui/material'
 
 import SafeSuspense from '@/src/components/helpers/SafeSuspense'
-import useIsBadgeOwner from '@/src/hooks/useIsBadgeOwner'
+import useIsBadgeOwner from '@/src/hooks/subgraph/useIsBadgeOwner'
 import AlreadyOwnThisBadgeError from '@/src/pagePartials/errors/displays/AlreadyOwnThisBadgeError'
 import { useWeb3Connection } from '@/src/providers/web3ConnectionProvider'
 import { ChainsValues } from '@/types/chains'
