+++ src/pagePartials/home/BadgesInReviewSwiper.tsx
@@ -8,11 +8,10 @@ import { A11y, Navigation } from 'swiper'
 import { Swiper, SwiperSlide } from 'swiper/react'
 
 import SafeSuspense from '@/src/components/helpers/SafeSuspense'
+import useSubgraph from '@/src/hooks/subgraph/useSubgraph'
 import { useSizeLG, useSizeMD, useSizeSM } from '@/src/hooks/useSize'
 import BadgeTypeMetadata from '@/src/pagePartials/badge/BadgeTypeMetadata'
 import { badgesExampleList } from '@/src/pagePartials/home/SectionBoxes'
-import { useWeb3Connection } from '@/src/providers/web3ConnectionProvider'
-import { SubgraphName, getSubgraphSdkByNetwork } from '@/src/subgraph/subgraph'
 
 const growEffect = keyframes`
   0% {
@@ -27,9 +26,8 @@ const now = Math.floor(Date.now() / 1000)
 export default function BadgesInReviewSwiper() {
   const theme = useTheme()
   const router = useRouter()
+  const gql = useSubgraph()
 
-  const { appChainId } = useWeb3Connection()
-  const gql = getSubgraphSdkByNetwork(appChainId, SubgraphName.TheBadge)
   const badgesInReview = gql.useBadgesInReview({ date: now })
   const sm = useSizeSM()
   const md = useSizeMD()
